Rename file to pill_number
